CRUD actions finished

+++ db/cinema.sql
@@ -5,13 +5,13 @@ DROP TABLE IF EXISTS films;
 CREATE TABLE films (
   id SERIAL PRIMARY KEY,
   title VARCHAR(255),
-  price MONEY
+  price INT
 );
 
 CREATE TABLE customers (
   id SERIAL PRIMARY KEY,
   name VARCHAR(255),
-  funds MONEY
+  funds INT
 );
 
 CREATE TABLE tickets (
